Added jar tasks for #16
