+++ hw07/src/main/kotlin/ru/hse/spb/client/RoguelikeClient.kt
@@ -124,29 +124,4 @@ internal constructor(private val channel: ManagedChannel) {
             }
         }
     }
-
-    companion object {
-        @Throws(Exception::class)
-        @JvmStatic
-        fun main(args: Array<String>) {
-            if (args.size > 2) {
-                printUsage()
-                return
-            }
-            if (args.size < 2) {
-                RoguelikeSinglePlayerApplication.start()
-                return
-            }
-            val client = RoguelikeClient(args[0], args[1].toInt())
-            try {
-                client.start()
-            } finally {
-                client.shutdown()
-            }
-        }
-
-        private fun printUsage() {
-            println("Args: <ip> <port> or empty. Last case is single player mode")
-        }
-    }
 }
